[gaitmesh_ros] Refactor GaitMesh into library
