+++ gaitmesh_ros/data/.gitignore
@@ -1,0 +1,3 @@
+*.ply
+*.obj
+*.dat
